refactor: move hooks
