+++ src/app/(main)/chat/(workspace)/features/ShareButton/index.tsx
@@ -7,9 +7,8 @@ import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { DESKTOP_HEADER_ICON_SIZE, MOBILE_HEADER_ICON_SIZE } from '@/const/layoutTokens';
+import { useWorkspaceModal } from '@/hooks/useWorkspaceModal';
 import { useChatStore } from '@/store/chat';
-
-import { useWorkspaceModal } from '../useWorkspaceModal';
 
 const ShareModal = dynamic(() => import('@/features/ShareModal'));
